refactor - integer only
requirement on the suggested interface was that should return an int

+++ Cameron.Katka.ClassLibrary/Models/Product.cs
@@ -4,9 +4,9 @@ namespace Cameron.Katka.ClassLibrary.Models
     public class Product
     {
         public string SKU { get; set; }
-        public decimal UnitPrice { get; set; }
+        public int UnitPrice { get; set; }
 
-        public Product(string sku, decimal unitPrice)
+        public Product(string sku, int unitPrice)
         {
             SKU = sku;
             UnitPrice = unitPrice;
